+++ assets/release_notes.md
@@ -9,6 +9,7 @@
 - Improvements for PartCategory editing form
 - Improvements for StockLocation editing form
 - Adds ability to edit StockItem
+- Display purchase price (where available) for StockItem
 - Updated translations 
 
 ### 0.2.10 - July 2021